+++ test.mjs
@@ -7,6 +7,7 @@ import test_cli_find_unused_exports from "./find-unused-exports.test.mjs";
 import test_findUnusedExports from "./findUnusedExports.test.mjs";
 import test_getVariableDeclarationIdentifierNames from "./getVariableDeclarationIdentifierNames.test.mjs";
 import test_isDirectoryPath from "./isDirectoryPath.test.mjs";
+import test_MODULE_GLOB from "./MODULE_GLOB.test.mjs";
 import test_reportCliError from "./reportCliError.test.mjs";
 import test_scanModuleCode from "./scanModuleCode.test.mjs";
 
@@ -17,6 +18,7 @@ test_cli_find_unused_exports(tests);
 test_findUnusedExports(tests);
 test_getVariableDeclarationIdentifierNames(tests);
 test_isDirectoryPath(tests);
+test_MODULE_GLOB(tests);
 test_reportCliError(tests);
 test_scanModuleCode(tests);